Add files via upload
This program reads the two .csv files that come with it. To add or remove accounts, you have to directly change the data in the csv file. The fields in the medicalstaff.csv file, separated by a comma, are userID,username,password,email,department. The fields in the patient.csv file, separated by a comma, are userID,username,password,email,treatmentInformation.

First off, you have the option to enter a username. Once you enter a username, you can enter a password. If you entered an incorrect username or password, you will be asked to reenter the two fields.

When you log in successfully as a patient using a correct username and password, you can view the data associated with that account and you can edit most of the corresponding fields. In addition, you can generate a couple of different reports on what data is found in the two files.

When you log in successfully as a staff member using a correct username and password, you can do everything a patient can, but you can also look up the data for specific patients and edit some of their data as well.